refactor: 리팩토링 및 Q클래스 경로 수동 수정
