+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/controller/AdminCommentController.java
@@ -1,7 +1,7 @@
 package com.sloweat.domain.admin.controller;
 
-import com.sloweat.domain.admin.dto.AdminCommentRequest;
-import com.sloweat.domain.admin.dto.AdminCommentResponse;
+import com.sloweat.domain.admin.dto.comment.AdminCommentRequest;
+import com.sloweat.domain.admin.dto.comment.AdminCommentResponse;
 import com.sloweat.domain.admin.service.AdminCommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
