+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/dto/comment/AdminCommentRequest.java
@@ -1,11 +1,11 @@
-package com.sloweat.domain.admin.dto;
+package com.sloweat.domain.admin.dto.comment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class AdminRecipeRequest {
-  private String title;
+public class AdminCommentRequest {
+  private String content;
   private String author;
 }
