+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/dto/comment/AdminCommentResponse.java
@@ -1,4 +1,4 @@
-package com.sloweat.domain.admin.dto;
+package com.sloweat.domain.admin.dto.comment;
 
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
