+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/dto/recipe/AdminRecipeRequest.java
@@ -1,11 +1,11 @@
-package com.sloweat.domain.admin.dto;
+package com.sloweat.domain.admin.dto.recipe;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class AdminCommentRequest {
-  private String content;
+public class AdminRecipeRequest {
+  private String title;
   private String author;
 }
