+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/repository/comment/AdminCommentRepository.java
@@ -1,4 +1,4 @@
-package com.sloweat.domain.admin.repository;
+package com.sloweat.domain.admin.repository.comment;
 
 import com.sloweat.domain.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
