+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/repository/comment/AdminCommentRepositoryCustom.java
@@ -1,7 +1,7 @@
-package com.sloweat.domain.admin.repository;
+package com.sloweat.domain.admin.repository.comment;
 
-import com.sloweat.domain.admin.dto.AdminCommentRequest;
-import com.sloweat.domain.admin.dto.AdminCommentResponse;
+import com.sloweat.domain.admin.dto.comment.AdminCommentRequest;
+import com.sloweat.domain.admin.dto.comment.AdminCommentResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
