+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/repository/recipe/AdminRecipeRepository.java
@@ -1,8 +1,7 @@
-package com.sloweat.domain.admin.repository;
+package com.sloweat.domain.admin.repository.recipe;
 
 import com.sloweat.domain.recipe.entity.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 // JPARepository는 기본 CRUD
 // AdminRecipeRepositoryCustom은 Querydsl 사용한 쿼리
